PULSEBS-104 patch lecture + small refactor whatWentWrong
